Added map_writer to main lib. Build and unit tests pass.
